+++ Demo/Demo/DemoKeys.swift
@@ -1,0 +1,18 @@
+//
+//  DemoKeys.swift
+//  Demo
+//
+//  Created by Daniel Saidi on 2023-03-28.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This struct is used to manage keys for this demo. Just type
+ in any keys you want to use. NEVER PUSH THESE KEYS!
+ */
+struct DemoKeys {
+
+    static var theMovieDb = "c9237ef2809ed01e64d7b37b0f951c7b"
+}