+++ Demo/Demo/DemoScreen.swift
@@ -1,0 +1,14 @@
+//
+//  DemoScreen.swift
+//  Demo
+//
+//  Created by Daniel Saidi on 2023-03-28.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+enum DemoScreen {
+
+    case theMovieDb
+}